A220 Extract SendableBuffer and ActorLifecycleEvent to own files
- actor_syscall/sendable_buffer.rs — zero-copy buffer transfer type
- actor_syscall/lifecycle_event.rs — actor lifecycle signals to executor
- io_bridge.rs imports them via super::
