+++ ailets-rs/ailetos/src/actor_syscall/mod.rs
@@ -1,10 +1,11 @@
+pub mod lifecycle_event;
 pub mod fd_table;
 pub mod io_bridge;
+pub mod sendable_buffer;
 pub mod stub_actor_runtime;
 
+pub use lifecycle_event::ActorLifecycleEvent;
 pub use fd_table::{FdEntry, FdTable};
-pub use io_bridge::{
-    ActorLifecycleEvent, Channel, ChannelHandle, IoBridge, IoEvent, IoFuture, IoRequest,
-    SendableBuffer,
-};
+pub use io_bridge::{Channel, ChannelHandle, IoBridge, IoEvent, IoFuture, IoRequest};
+pub use sendable_buffer::SendableBuffer;
 pub use stub_actor_runtime::{BlockingActorRuntime, ShutdownHandle};
